requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile ./requirements.in
#
absl-py==2.1.0
    # via
    #   mujoco
    #   mujoco-mjx
etils[epath,epy]==1.9.4
    # via
    #   mujoco
    #   mujoco-mjx
fsspec==2024.9.0
    # via etils
glfw==2.7.0
    # via mujoco
importlib-resources==6.4.5
    # via etils
jax==0.4.33
    # via mujoco-mjx
jaxlib==0.4.33
    # via
    #   jax
    #   mujoco-mjx
ml-dtypes==0.5.0
    # via
    #   jax
    #   jaxlib
mujoco==3.2.3
    # via
    #   -r ./requirements.in
    #   mujoco-mjx
mujoco-mjx==3.2.3
    # via -r ./requirements.in
numpy==2.1.1
    # via
    #   jax
    #   jaxlib
    #   ml-dtypes
    #   mujoco
    #   scipy
    #   trimesh
opt-einsum==3.4.0
    # via jax
pyopengl==3.1.7
    # via mujoco
scipy==1.14.1
    # via
    #   jax
    #   jaxlib
    #   mujoco-mjx
trimesh==4.4.9
    # via mujoco-mjx
typing-extensions==4.12.2
    # via etils
zipp==3.20.2
    # via etils
